chore: refactor and add tests
