+++ packages/cli/src/__mocks__/fs.ts
@@ -2,3 +2,4 @@ export const existsSync = jest.fn();
 export const readFileSync = jest.fn(() => '');
 export const statSync = jest.fn(() => ({ size: 0 }));
 export const createReadStream = jest.fn();
+export const writeFileSync = jest.fn();
