+++ packages/cli/src/__mocks__/redoc.ts
@@ -1,0 +1,2 @@
+export const loadAndBundleSpec = jest.fn(() => Promise.resolve({ openapi: '3.0.0' }));
+export const createStore = jest.fn(() => Promise.resolve({ toJS: jest.fn(() => '{}') }));
